Fixed MariaDB password details
The instructions told the user to  replace all instances of "t00r" in certain
parts of the code with the Tomcat password they choose.
This was problematic because it should have instead said MariaDB
password as that is what the code was expecting. Also, it's better to
make the user pick "t00r" as the password for MariaDB because
that is what Osbert is using (as last known at the time of this commit)
and will make it easier for students to share new versions of the
project with him.